cahnges in script for run test and changes y file for testings
