+++ tests/add.media.type.tolist.test.ts
@@ -1,52 +1,17 @@
 import {server} from '../src'
-// @ts-ignore
-import {
-    generatedResponseAddMovieToWatchList,
-    generateExceptionWhenAddFavoriteMovie, generateExceptionWhenAddWatchList,
-    generateFakeDetailsAccount,
-    generateMovieInfoToAdd,
-    generateMovieInfoToAddUpdate, generateResponseRemoveFromFavorites, generateResponseRemoveFromWatchlist
-} from "./utilities/AccounDataFake";
-// @ts-ignore
 import {
     addAMovieToWatchList,
     deleteMovieFromFavoritesMovies,
     errorInServiceAddFavoriteMovie, errorInServiceAddWatchlist,
     generateAddFavoriteMovie, removeMovieFromWatchlist
 } from "./utilities/RequestData";
-import exp = require("constants");
-
-const {api} = require('./helpers')
-
-describe('TMDB API get Account Description test', () => {
-    it('fetches account details successfully', async () => {
-        const expectedData = generateFakeDetailsAccount()
-
-        const response = await api
-            .get('/api/account/details/10332765')
-            .expect(200)
-
-        expect(response.status).toBe(200)
-        expect(response.body).toEqual(expectedData)
-    })
-
-    it('is not possible the accountId because url is note correct', async () => {
-        const response = await api.get('/api/account/details/').expect(404)
-
-        expect(response.status).toBe(404)
-    })
-
-    it('Is not possible the accountId is not valid', async () => {
-        const requestError = {
-            codeMessage: 'TMDB0001',
-            message: 'AccountId is required and must be an integer.'
-        }
-        const response = await api.get('/api/account/details/fr2233o').expect(400)
-
-        expect(response.status).toBe(400)
-        expect(response.body).toEqual(requestError)
-    })
-})
+import {
+    generatedResponseAddMovieToWatchList,
+    generateExceptionWhenAddFavoriteMovie, generateExceptionWhenAddWatchList,
+    generateMovieInfoToAdd,
+    generateMovieInfoToAddUpdate, generateResponseRemoveFromFavorites, generateResponseRemoveFromWatchlist
+} from "./utilities/AccounDataFake";
+import {api} from "./helpers";
 
 describe('TMDB add Movie to favorite list', () => {
     it('Add new Movie to the list', async () => {
@@ -141,7 +106,7 @@ describe('Add a movie to the list of Watchlist ', () => {
         expect(response.body).toEqual(responseError)
     })
 
-    it('Is not possible ', async () => {
+    it('Is not possible to add, send incorrect type of media type', async () => {
         const requestObject = errorInServiceAddWatchlist()
         const responseObject = generateExceptionWhenAddWatchList()
 
@@ -169,5 +134,4 @@ describe('Add a movie to the list of Watchlist ', () => {
 
 afterAll(() => {
     server.close()
-    api.release
 })
