+++ tsconfig.json
@@ -19,6 +19,6 @@
   ],
   "exclude": [
     "node_modules",                            // Exclude the node_modules directory.
-    "**/*.test.ts"                             // Exclude test files.
+    "**/*.test.ts"                             // Exclude tests files.
   ]
 }
